Enable the unique_id for GPU devices

+++ common/src/resource.rs
@@ -23,7 +23,7 @@ pub struct ResourceReq {
 #[derive(Clone, Debug, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
 pub struct ResourceAlloc {
     pub requirement: ResourceReq,
-    pub resource_id: Vec<usize>,
+    pub resource_id: Vec<u64>,
 }
 
 impl Default for ResourceAlloc {
