+++ scheduler/src/server.rs
@@ -9,7 +9,7 @@ use crate::requests::{SchedulerRequest, SchedulerResponse};
 use crate::Error;
 use common::{ClientToken, RequestMethod, ResourceAlloc, TaskRequirements};
 
-type AllocationResult = std::result::Result<Vec<(usize, u64)>, Error>;
+type AllocationResult = std::result::Result<Vec<(u64, u64)>, Error>;
 
 #[rpc(server)]
 pub trait RpcMethods {
